requirements.txt
scikit-learn==1.1.3
pandas==1.5.3
numpy==1.23.5
matplotlib==3.4.3
joblib==1.1.0
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==1.1.0
threadpoolctl==3.6.0
scipy==1.15.3
sqlalchemy==2.0.30
aiosqlite==0.20.0
requests
